Parser weiter ausgebaut

+++ Routenplaner/main/de/dhbw/horb/routePlanner/graphData/Node.java
@@ -3,30 +3,30 @@ package de.dhbw.horb.routePlanner.graphData;
 public class Node {
 
 	private String name;
-	private float latitude;
-	private float longitude;
+	private Double latitude;
+	private Double longitude;
 	
 	
-	public Node(String name, float latitude, float longitude) {
+	public Node(String name, Double latitude, Double longitude) {
 		
 		setName(name);
 		setLatitude(latitude);
 		setLongitude(longitude);
 	}
 	
-	public float getLatitude() {
+	public Double getLatitude() {
 		return latitude;
 	}
 
-	public void setLatitude(float latitude) {
+	public void setLatitude(Double latitude) {
 		this.latitude = latitude;
 	}
 
-	public float getLongitude() {
+	public Double getLongitude() {
 		return longitude;
 	}
 
-	public void setLongitude(float longitude) {
+	public void setLongitude(Double longitude) {
 		this.longitude = longitude;
 	}
